Added support for backend secret key

+++ ttg-website/start-ttg-website.sh
@@ -6,7 +6,15 @@ then
     >&2 printf " -> Missing MongoDB home volume argument (arg 1)\n"
     exit 1
 fi
+
+if [ -z "${TTG_WEB_BACKEND_SECRET}" ]; then
+   >&2 printf " -> Missing TTG_WEB_BACKEND_SECRET variable definition"
+   exit 1
+fi
+
+export TTG_WEB_BACKEND_SECRET="${TTG_WEB_BACKEND_SECRET}"
 export TTG_MONGO_VOL="${1}"
+
 printf " -> Destroying any existing ttg-web containers... \n\n"
 docker-compose down
 printf " -> Done!\n"
